P4PU-358: Centralized pre login layout
